git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/solr2452@1144174 13f79535-47bb-0310-9956-ffa450edef68

+++ solr/core/src/java/org/apache/solr/request/SimpleFacets.java
@@ -1317,7 +1317,7 @@ public class SimpleFacets {
     }
     @Override
     public Double parseAndAddGap(Double value, String gap) {
-      return new Double(value.floatValue() + Double.valueOf(gap).floatValue());
+      return new Double(value.doubleValue() + Double.valueOf(gap).doubleValue());
     }
   }
   private static class IntegerRangeEndpointCalculator 
@@ -1343,7 +1343,7 @@ public class SimpleFacets {
     }
     @Override
     public Long parseAndAddGap(Long value, String gap) {
-      return new Long(value.intValue() + Long.valueOf(gap).intValue());
+      return new Long(value.longValue() + Long.valueOf(gap).longValue());
     }
   }
   private static class DateRangeEndpointCalculator
